requirements.txt
pytest
requests
pytest-html
pandas
openpyxl
python-docx
pytest-playwright
